Issues with User model fixed Role model separated
