+++ app/model/role.py
@@ -1,0 +1,8 @@
+import enum
+
+
+class Role(enum.Enum):
+    Admin = "Admin"
+    Owner = "Owner"
+    Client = "Client"
+    NoneUser = "None"
